ftl/requirements.txt
django==2.2.9
django-mptt==0.9.1
django-webpack-loader==0.6.0
psycopg2==2.8.3
djangorestframework==3.10.3
tika==1.19
langid==1.1.6
whitenoise==4.1.3
djangorestframework_simplejwt==4.3.0
filetype==1.0.5
django-registration==3.0.1
django-otp==0.7.2
fido2==0.7.1
cbor2==4.1.2
qrcode[pil]==6.1

# django-storages==1.7.1
# boto3==1.9.183
# google-cloud-storage==1.18.0
# google-cloud-vision==0.38.1
